refactor(restaurant-detail): reviewItem 파일 이름 변경
